app/requirements.txt
django==5.0
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
django-environ==0.11.2
drf-spectacular==0.27.2
channels==4.1.0
daphne==4.1.2
channels_redis==4.2.0
django-cors-headers==4.4.0
psycopg2-binary~=2.0
gunicorn==22.0.0
